requirements.txt
asgiref==3.8.1
chardet==5.2.0
cssselect2==0.7.0
Django==5.0.7
django-ckeditor==6.7.1
django-filer==3.2.3
django-filter==24.3
django-js-asset==2.2.0
django-polymorphic==3.1.0
djangorestframework==3.15.2
easy-thumbnails==2.10
lxml==5.3.0
Markdown==3.6
pillow==10.4.0
reportlab==4.2.5
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.3.0
webencodings==0.5.1
